# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: thelake API
  version: 0.1.0
  description: |
    Authenticated HTTP API exposed by thelake. Interactive UI: `GET /swagger`
    (this document is also served as `GET /openapi.json`).

    This API accepts telemetry payloads on standard OTLP HTTP paths:
    - `POST /v1/traces`
    - `POST /v1/logs`

    It also documents the tenant-scoped schema promotion endpoint:
    - `POST /v1/promotions/apply`

    Evaluations are written through:
    - `POST /v1/scores`
    - `GET|POST /v1/score-configs`

    Query surfaces:
    - `POST /v1/traces/search`
    - `POST /v1/traces/details`
    - `POST /v1/spans/search`
    - `GET /v1/spans/{span_id}`
    - `GET /v1/traces/{trace_id}`
    - `POST /v1/sessions/search`
    - `POST /v1/sessions/details`
    - `POST /v1/sessions/summary/rebuild`
    - `GET /v1/sessions/{session_id}`
    - `GET /v1/sessions/{session_id}/recording`

    Control plane:
    - `POST /v1/workspaces` (admin bearer)
    - `GET /v1/meta`
    - `POST /v1/promotions/apply`
    - `GET /v1/data/ducklake-connection`

    Also inventoried here (summary-level): health/ready, `POST /v1/query/sql`,
    and `/v1/fields`. Loki/Tempo Grafana-compat query routes are served by
    the process but are not part of this product OpenAPI document.

    For JSON payloads, requests are expected to follow the OpenTelemetry collector export
    request shapes (`ExportTraceServiceRequest` and `ExportLogsServiceRequest`).

...

...

...

...
    `docs/how-to/promotion.md`.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/traces:
    post:
      tags:
        - ingestion
      summary: Ingest OTLP traces
      operationId: ingestTraces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportTraceServiceRequest"
          application/x-protobuf:
            schema:
              type: string
              format: binary
              description: Protobuf-encoded `ExportTraceServiceRequest`
      responses:
        "200":
          description: Request accepted and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          description: Invalid JSON/protobuf payload
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                invalidJson:
                  value: Invalid JSON
                invalidProtobuf:
                  value: Protobuf decode failed
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "413":
          description: Request exceeds `server.max_body_size`
        "503":
          $ref: "#/components/responses/IngestUnavailable"
  /v1/logs:
    post:
      tags:
        - ingestion
      summary: Ingest OTLP logs
      operationId: ingestLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportLogsServiceRequest"
          application/x-protobuf:
            schema:
              type: string
              format: binary
              description: Protobuf-encoded `ExportLogsServiceRequest`
      responses:
        "200":
          description: Request accepted and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          description: Invalid JSON/protobuf payload
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                invalidJson:
                  value: "Invalid JSON: expected value at line 1 column 1"
                invalidProtobuf:
                  value: Protobuf decode failed
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "413":
          description: Request exceeds `server.max_body_size`
        "503":
          $ref: "#/components/responses/IngestUnavailable"
  /v1/scores:
    post:
      tags:
        - llm
      summary: Create an immutable LLM evaluation score
      description: |
        Creates a score in the authenticated tenant's DuckLake scope. Tenant
        identity is resolved from the bearer token and is never accepted in
        the request body.

        At least one target (`trace_id`, `span_id`, or `session_id`) is
        required. Exactly one value field must be present and must match
        `data_type`:

        - `numeric` uses `numeric_value`
        - `categorical` and `text` use `string_value`
        - `boolean` uses `boolean_value`

        `score_id` is the tenant-local idempotency key. Repeating a previously
        accepted ID does not append another score row.
      operationId: createScore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoreRequest"
            examples:
              correctness:
                value:
                  score_id: score-01JZ8YH3KQBZ8V2CA9ZB1W7Q9A
                  timestamp: "2026-07-18T23:22:00Z"
                  trace_id: 4bf92f3577b34da6a3ce929d0e0e4736
                  span_id: 00f067aa0ba902b7
                  name: correctness
                  data_type: numeric
                  numeric_value: 0.92
                  source: evaluator
                  comment: Answer is supported by the retrieved context.
                  metadata:
                    evaluator: groundedness-v2
      responses:
        "201":
          description: Score created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "200":
          description: Idempotent retry; the score ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "400":
          description: Missing target, missing value, or value does not match `data_type`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "422":
          description: Malformed JSON or field type mismatch
        "503":
          description: Tenant runtime, idempotency lookup, or DuckLake write unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/policies:
    get:
      tags: [llm]
      summary: List workspace and agent Markdown policy versions
      description: Returns workspace-scoped immutable Markdown policy versions.
      operationId: listPolicies
      responses:
        "200":
          description: Policy documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyDocument"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
    post:
      tags: [llm]
      summary: Create an immutable Markdown policy version
      description: Stores workspace-wide or exact-agent policy content in the authenticated workspace.
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDocument"
        "200":
          description: Idempotent retry of the same policy version
        "400":
          description: Invalid policy or conflicting immutable version
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
  /v1/evaluators:
    get:
      tags: [llm]
      summary: List no-code behavior evaluators
      description: Returns versioned evaluator criteria stored in the authenticated workspace.
      operationId: listEvaluators
      responses:
        "200":
          description: Evaluator definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvaluatorDefinition"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
    post:
      tags: [llm]
      summary: Create a draft no-code behavior evaluator version
      description: |
        Stores a natural-language criterion and optional ordered tool
        requirement for one named agent in the authenticated workspace. Drafts
        do not trigger evaluations. Explicit activation opts this evaluator
        into sending matching trace content to the configured judge provider.
      operationId: createEvaluator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvaluatorRequest"
            examples:
              eligibilityBeforeRefund:
                value:
                  evaluator_id: refund-eligibility
                  version: 1
                  target_agent_name: support-agent
                  name: Refund eligibility before refund
                  criteria: Verify the agent checks refund eligibility and explains the result before issuing a refund.
                  threshold: 0.7
                  uncertainty_margin: 0.1
                  required_tool_order:
                    - before: check_eligibility
                      action: issue_refund
                      require_result_before_action: true
      responses:
        "201":
          description: Evaluator version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluatorDefinition"
        "200":
          description: Idempotent retry of the same evaluator version
        "400":
          description: Invalid evaluator definition or conflicting config ID
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
...
  /v1/score-configs:
    get:
      tags:
        - llm
      summary: List score configs (seeds defaults when empty)
      description: |
        Returns append-only score schemas for the authenticated tenant. When the
        tenant has no configs yet, the server seeds built-in defaults
        (`correctness`, `quality`, `expected_output`) and returns them.
      operationId: listScoreConfigs
      responses:
        "200":
          description: Score config list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreConfigListResponse"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "503":
          description: Tenant runtime or DuckLake unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
        - llm
      summary: Create an append-only score config
      description: |
        Creates a score config in the authenticated tenant's DuckLake scope.
        `config_id` is the tenant-local idempotency key.
      operationId: createScoreConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoreConfigRequest"
            examples:
              correctness:
                value:
                  config_id: cfg-correctness
                  timestamp: "2026-07-18T23:22:00Z"
                  name: correctness
                  data_type: boolean
                  description: Whether the turn was correct
                  author_id: system
      responses:
        "201":
          description: Score config created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreConfig"
        "200":
          description: Existing config returned after an idempotent retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreConfig"
        "400":
          description: Invalid score config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "503":
          description: Tenant runtime or DuckLake write unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/spans/search:
    post:
      tags:
        - llm
      summary: Search lightweight LLM span projections
      description: |
        Requires `from` and `to`. Returns lightweight span rows without
        full attributes/events. LLM fields are derived from span attributes
        (`gen_ai.*`, `sp.*`) and do not require promotion to be applied.
        Cursor ordering is `(timestamp DESC, span_id DESC)`.
      operationId: searchSpans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpanSearchRequest"
      responses:
        "200":
          description: Bounded span page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpanSearchResponse"
        "400":
          description: Missing time range, invalid filter, or malformed cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "503":
          description: Tenant runtime or query unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /v1/sessions/search:
    post:
      tags:
        - llm
      summary: Search sessions via session_summary
      description: |
        Always backed by Postgres `session_summary` (empty table → empty page).
        Requires `from` and `to`. Cursor paging is only defined for
        `order_by=start_time`; other orderings return a single ranked page and
        set `cursor_supported=false`.
      operationId: searchSessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionSearchRequest"
      responses:
        "200":
          description: Bounded session page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSearchResponse"
        "400":
          description: Missing time range, invalid filter, or malformed cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "403":
          description: Bearer token could not be resolved to a tenant
        "503":
          description: session_summary or tenant runtime unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /v1/traces/search:
    post:
      tags: [llm]
      summary: Search traces with a dynamic filter expression
      description: |
        Uses `TelemetrySearchRequest`: `version`, `scope` (`traces`), required
        nested `timeRange` (`from`/`to`), optional `filter` / `columns` / `sort` /
        `limit` / `cursor`. JSON fields are camelCase.
      operationId: searchTraces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetrySearchRequest"
      responses:
        "200":
          description: Bounded trace search page
        "400":
          description: Missing timeRange, invalid filter, or malformed cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid bearer token
        "503":
          description: Workspace runtime or query unavailable
...
  /v1/workspaces:
    post:
      tags: [control]
      summary: Provision a workspace DuckLake scope
      description: |
        Admin-only. Requires `Authorization: Bearer` matching
        `SOFTPROBE_ADMIN_API_KEY`. Body uses camelCase `workspaceId` and
        required `storageHints` (`ducklakeMetadataSchema`, `ducklakeDataPath`).
      operationId: provisionWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, storageHints]
              properties:
                workspaceId:
                  type: string
                  format: uuid
                storageHints:
                  type: object
                  required: [ducklakeMetadataSchema, ducklakeDataPath]
                  properties:
                    ducklakeMetadataSchema:
                      type: string
                    ducklakeDataPath:
                      type: string
                    gcsBucket:
                      type: string
      responses:
        "200":
          description: Scope created or already exists with matching hints
        "400":
          description: Invalid workspace id, reserved id, or missing storage hints
        "401":
          description: Missing bearer token
        "403":
          description: Admin API key required
        "409":
          description: Workspace exists with a different storage scope
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    IngestUnavailable:
      description: DuckLake commit failed after backend retries; exporter should retry
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IngestResponse"
          example:
            success: false
            ingested_count: 0
            message: "Ingestion failed: DuckLake write failed"
  schemas:
...
    CreatePolicyRequest:
      type: object
      required: [policy_id, version, content]
      properties:
        policy_id:
          type: string
          maxLength: 200
          description: Stable policy ID, unique within its workspace.
        version:
          type: integer
          minimum: 1
        target_agent_name:
          type: string
          nullable: true
          maxLength: 256
          description: Exact agent name recorded on telemetry; omit for workspace-wide policy.
        content:
          type: string
          minLength: 1
          maxLength: 32000
          description: Human-readable Markdown POLICY.md content.
    PolicyDocument:
      allOf:
        - $ref: "#/components/schemas/CreatePolicyRequest"
        - type: object
          required: [timestamp]
          properties:
            timestamp:
              type: string
              format: date-time
...
    EvaluatorDefinition:
      allOf:
        - $ref: "#/components/schemas/CreateEvaluatorRequest"
        - type: object
          required: [timestamp, active]
          properties:
            timestamp:
              type: string
              format: date-time
            active:
              type: boolean
    IngestResponse:
      type: object
      required:
        - success
        - ingested_count
        - message
      properties:
        success:
          type: boolean
          description: True when ingestion processing succeeded.
        ingested_count:
          type: integer
          minimum: 0
          description: Number of telemetry records successfully ingested.
        message:
          type: string
          description: Human-readable result details.
      examples:
        traces:
          value:
            success: true
            ingested_count: 128
            message: Successfully ingested 128 spans
    ErrorMessage:
      type: string
      description: Plain-text error body returned for malformed requests.
    ExportTraceServiceRequest:
      type: object
      description: |
        OTLP trace export request (JSON mapping).
        See OpenTelemetry protocol spec for the canonical schema.
      additionalProperties: true
    ExportLogsServiceRequest:
      type: object
      description: |
        OTLP logs export request (JSON mapping).
        See OpenTelemetry protocol spec for the canonical schema.
      additionalProperties: true
    CreateScoreRequest:
      type: object
      required:
        - score_id
        - timestamp
        - name
        - data_type
        - source
      description: |
        At least one target ID and exactly one value matching `data_type` are
        required. Tenant identity is intentionally absent.
      properties:
        score_id:
          type: string
          minLength: 1
        timestamp:
          type: string
          format: date-time
        trace_id:
          type: string
        span_id:
          type: string
        session_id:
          type: string
        name:
          type: string
          minLength: 1
        data_type:
          type: string
          enum: [numeric, categorical, boolean, text]
        numeric_value:
          type: number
        string_value:
          type: string
        boolean_value:
          type: boolean
        source:
          type: string
          enum: [api, user, evaluator, annotation]
        comment:
          type: string
        config_id:
          type: string
        author_id:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
          default: {}
    Score:
      allOf:
        - $ref: "#/components/schemas/CreateScoreRequest"
        - type: object
          required:
            - timestamp
          properties:
            timestamp:
              type: string
              format: date-time
...
    ScoreConfig:
      allOf:
        - $ref: "#/components/schemas/CreateScoreConfigRequest"
        - type: object
          required:
            - timestamp
          properties:
            timestamp:
              type: string
              format: date-time
    ScoreConfigListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScoreConfig"
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...